requirements.txt
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0
2captcha-python>=1.3
